Move rust memchr impl to libcore
