+++ src/libstd/sys_common/mod.rs
@@ -33,7 +33,6 @@ pub mod at_exit_imp;
 pub mod backtrace;
 pub mod condvar;
 pub mod io;
-pub mod memchr;
 pub mod mutex;
 pub mod poison;
 pub mod remutex;
